Refactor calendar and growth tabs
