+++ app/(tabs)/growth/index.tsx
@@ -1,2 +1,3 @@
+//C:\Users\fukur\task-app\app\app\(tabs)\growth\index.tsx
 import GrowthScreen from '@/features/growth';
 export default GrowthScreen;
