+++ features/calendar/index.tsx
@@ -1,0 +1,2 @@
+import CalendarScreen from './CalendarScreen';
+export default CalendarScreen;
